services/ingestion-service/requirements.txt
# Ingestion Service Dependencies

# gRPC
grpcio>=1.62.0  # protobuf 5.x compatibility
grpcio-tools>=1.62.0
protobuf>=5.29.2  # Security update

# Message Queue
pika==1.3.2

# Async
asyncio==3.4.3

# HTTP client (for HTTP protocol)
httpx==0.25.2

# Logging
structlog==24.1.0

# Configuration
python-dotenv==1.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
PyYAML==6.0.1
cryptography>=42.0.4  # CVE-2024-26130, CVE-2023-50782 fix

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

